SWD-26 generate mapper for every dto

+++ backend/src/main/java/com/fptgang/backend/repository/NotificationRepos.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface NotificationRepos extends JpaSpecificationExecutor<Notification>, JpaRepository<Notification, Integer> {
+public interface NotificationRepos extends JpaSpecificationExecutor<Notification>, JpaRepository<Notification, Long> {
 }
